changing syntax highlighter to Google prettify
