+++ src/main/scala/edu/ucsc/vesper/http/core/Interpreter.scala
@@ -2,19 +2,17 @@ package edu.ucsc.vesper.http.core
 
 import edu.ucsc.refactor._
 import edu.ucsc.refactor.spi._
-import edu.ucsc.refactor.util.{SourceFormatter, Commit}
+import edu.ucsc.refactor.util.Commit
 import edu.ucsc.vesper.http.config.Configuration
 import edu.ucsc.vesper.http.domain.Models._
-import edu.ucsc.vesper.http.util.{LinkifyTweet, GoogleUrlShortener}
-import twitter4j.{Twitter, Status, TwitterFactory}
+import edu.ucsc.vesper.http.util.{GoogleUrlShortener, Html}
 import twitter4j.conf.ConfigurationBuilder
+import twitter4j.{Status, Twitter, TwitterFactory}
 
 import scala.collection.JavaConversions._
 import scala.collection.mutable
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.Try
-import scalatags.Text.all._
-import scalatags.Text.tags2.title
 
 /**
  * @author hsanchez@cs.ucsc.edu (Huascar A. Sanchez)
@@ -613,64 +611,8 @@ trait Interpreter extends Configuration with VesperConversions with CommandFlatt
     }
   }
 
-  def linkify(text: String) = {
-    LinkifyTweet.createLinks(text)
-  }
-
   private def page(theCode: Code) = {
-    val hrefUrl   = theCode.url.getOrElse("#")
-    val txtUrl    = theCode.url.getOrElse("(Missing URL)")
-    val linkified = linkify(theCode.description.stripSuffix(".") + " " +
-      List.concat(
-        theCode.tags,
-        theCode.algorithms,
-        theCode.datastructures
-      ).distinct.map(x => "#" + x).mkString(" "))
-
-    val codesnippet = html(
-      head(
-        meta(charset := "utf-8"),
-        title(theCode.name),
-
-        /* css links */
-        link(rel:= "stylesheet", href:="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css"),
-        link(rel:= "stylesheet", href:="http://yandex.st/highlightjs/8.0/styles/tomorrow.min.css")
-      ),
-
-      body(
-        // main
-        div(`class`:="banner well", style:="background: #ffffff; border: none")(
-          h3(theCode.name + " ", a(href:= theCode.url.getOrElse(hrefUrl))(txtUrl)),
-          p(raw(linkified)),
-          // code
-          div(
-            pre(style:= "display: block; padding: 6px 4px; background: #fff; border: 1px dashed #ddd; margin-top:5px;") (
-              code(`class`:="java", style:="font-size: 12px;")(
-                raw(new SourceFormatter().format(theCode.content))
-              )
-            )
-          ), // end of code
-          // footer
-          div(id:="footer_wrap", `class`:= "outer")(
-            footer(`class`:= "inner")(
-              p(id:= "project_copyright", `class`:= "copyright")(
-                "HTML was generated by ",
-                strong()("Vesper"),
-                ", a tool maintained by ",
-                a(href:="https://github.com/hsanchez")("Huascar A. Sanchez")
-              )
-            )
-          ) // end of footer
-        ), // end of main
-        script(src:="https://code.jquery.com/jquery-2.1.1.js"),
-        script(src:="http://yandex.st/highlightjs/8.0/highlight.min.js"),
-        script(src:="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"),
-        script("hljs.initHighlightingOnLoad();")
-      )
-
-    )
-
-    scala.xml.Unparsed(codesnippet.toString())
+    Html(theCode).createHtmlElement()
   }
 
 }
