+++ src/main/scala/edu/ucsc/vesper/http/util/Html.scala
@@ -1,0 +1,84 @@
+package edu.ucsc.vesper.http.util
+
+import edu.ucsc.refactor.util.SourceFormatter
+import edu.ucsc.vesper.http.domain.Models.Code
+
+import scalatags.Text.all._
+import scalatags.Text.tags2.title
+
+/**
+ * @author hsanchez@cs.ucsc.edu (Huascar A. Sanchez)
+ */
+case class Html(theCode: Code) {
+
+  private def linkify(text: String) = {
+    LinkifyText.createLinks(text)
+  }
+
+  def createHtmlElement() = {
+    val hrefUrl   = theCode.url.getOrElse("#")
+    val txtUrl    = theCode.url.getOrElse("(Missing URL)")
+    val linkified = linkify(theCode.description.stripSuffix(".") + " " +
+      List.concat(
+        theCode.tags,
+        theCode.algorithms,
+        theCode.datastructures
+      ).distinct.map(x => "#" + x).mkString(" "))
+
+    val codesnippet = html(
+      head(
+        meta(charset := "utf-8"),
+        title(theCode.name),
+
+        /* css links */
+        link(rel:= "stylesheet", href:="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css"),
+//        link(rel:= "stylesheet", href:="http://yandex.st/highlightjs/8.0/styles/tomorrow.min.css")
+        link(rel:= "stylesheet", href:="http://google-code-prettify.googlecode.com/svn/trunk/src/prettify.css"),
+        link(rel:= "stylesheet", href:="http://demo.stanleyhlng.com/prettify-js/assets/themes/tomorrow-night-light/prettify.css")
+      ),
+
+      body(
+        // main
+        div(`class`:="banner well", style:="background: #ffffff; border: none")(
+          h3(theCode.name + " ", a(href:= theCode.url.getOrElse(hrefUrl))(txtUrl)),
+          p(raw(linkified)),
+          // code
+          div(
+            pre(style:= "display: block; padding: 6px 4px; background: #fff; border: 1px dashed #ddd; margin-top:5px;") (
+              code(`class`:="prettyprint linenums lang-java", style:="font-size: 12px;")(
+                raw(new SourceFormatter().format(theCode.content))
+              )
+            )
+          ), // end of code
+          // footer
+          div(id:="footer_wrap", `class`:= "outer")(
+            footer(`class`:= "inner")(
+              p(id:= "project_copyright", `class`:= "copyright")(
+                "HTML was generated by ",
+                strong()("Vesper"),
+                ", a tool maintained by ",
+                a(href:="https://github.com/hsanchez")("Huascar A. Sanchez")
+              )
+            )
+          ) // end of footer
+        ), // end of main
+        script(src:="https://code.jquery.com/jquery-2.1.1.js"),
+//        script(src:="http://yandex.st/highlightjs/8.0/highlight.min.js"),
+        script(src:="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"),
+        script(src:="http://google-code-prettify.googlecode.com/svn/trunk/src/prettify.js"),
+//        script("hljs.initHighlightingOnLoad();")
+        script(
+          """
+            | (function(jQuery){
+            |   jQuery( document ).ready( function() {
+            |     prettyPrint();
+            |   });
+            | }(window.jQuery))
+          """.stripMargin)
+      )
+
+    )
+
+    scala.xml.Unparsed(codesnippet.toString())
+  }
+}
